made some progress on moving files
