+++ src/ce/delete_edge.py
