+++ src/ce/delete_glb.py
